requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
attrs==25.3.0
cfgv==3.4.0
contourpy==1.3.2
cycler==0.12.1
distlib==0.3.9
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
gensim==4.3.3
hydra-core==1.3.2
identify==2.6.10
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
lightning-utilities==0.14.3
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
omegaconf==2.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
propcache==0.3.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.13.1
setuptools==78.1.1
six==1.17.0
smart-open==7.1.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
torchmetrics==1.7.1
torchsummary==1.5.1
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
virtualenv==20.30.0
wrapt==1.17.2
yarl==1.20.0
